+++ src/tracing/instrumentation/httpServer.js
@@ -5,7 +5,6 @@ var coreHttpModule = require('http');
 
 var discardUrlParameters = require('../../util/url').discardUrlParameters;
 var tracingConstants = require('../constants');
-var transmission = require('../transmission');
 var shimmer = require('shimmer');
 var cls = require('../cls');
 
@@ -35,8 +34,7 @@ function shimEmit(realEmit) {
 
       var incomingTraceId = getExistingTraceId(req);
       var incomingSpanId = getExistingSpanId(req);
-      // expose the span for the express instrumentation
-      var span = req.__instanaHttpSpan = cls.startSpan(exports.spanName, incomingTraceId, incomingSpanId);
+      var span = cls.startSpan(exports.spanName, incomingTraceId, incomingSpanId);
 
       // Grab the URL before application code gets access to the incoming message.
       // We are doing this because libs like express are manipulating req.url when
@@ -79,7 +77,7 @@ function shimEmit(realEmit) {
         span.error = res.statusCode >= 500;
         span.ec = span.error ? 1 : 0;
         span.d = Date.now() - span.ts;
-        transmission.addSpan(span);
+        span.transmit();
       });
 
       cls.ns.bindEmitter(req);